experiments/image_seg_lottery/requirements.txt
tqdm
zipfile36
scikit-image
scikit-learn
tensorflow-gpu==1.14